track: Create pokemon (7)

+++ configure.sh
@@ -99,8 +99,15 @@ if [ -d "$source" ]; then
 
     path="$PWD/track/$1/dist/content"
 
-    cp "$source"/*.png "$path"
-    cp "$source"/*.mp4 "$path"
+    if exists "$source"/*.png; then
+
+        cp "$source"/*.png "$path"
+    fi
+
+    if exists "$source"/*.mp4; then
+
+        cp "$source"/*.mp4 "$path"
+    fi
 fi
 
 source="$assets/track/$1/dist/content/upscale"
